refactor: Change id type to long
- 더 큰 범위를 받을 수 있는 long타입으로 고치고 id의 위치를 최상단으로 바꿨습니다.

+++ src/main/java/com/codessquad/qna/controller/QnaController.java
@@ -29,8 +29,8 @@ public class QnaController {
     }
 
     @GetMapping("/questions/{id}")
-    public String getDetailedQuestion(@PathVariable int id, Model model) {
-        model.addAttribute("question", questions.get(id - 1));
+    public String getDetailedQuestion(@PathVariable long id, Model model) {
+        model.addAttribute("question", questions.get((int) (id - 1)));
         return "qna/show";
     }
 }
